requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.10
async-timeout==5.0.1
attrs==25.3.0
bcrypt==4.3.0
bidict==0.23.1
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
dill==0.4.0
dnspython==2.7.0
docutils==0.21.2
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
isodate==0.7.2
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
motor==3.7.0
multidict==6.4.3
mypy==1.15.0
mypy_extensions==1.1.0
nicegui==2.17.0
numpy==1.24.4
openapi-python-client==0.24.3
orjson==3.10.18
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.8
propcache==0.3.1
pscript==0.7.7
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pylint==3.3.7
pymongo==4.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.1
python-multipart==0.0.20
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
ruff==0.11.9
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tomli==2.2.1
tomlkit==0.13.2
typer==0.15.3
types-passlib==1.7.7.20250408
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
vbuild==0.8.2
wait-for2==0.3.2
watchfiles==1.0.5
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.0
redis==6.2.0
notion-client==2.4.0
